+++ optd-dsl/src/engine/test_utils.rs
@@ -2,7 +2,7 @@ use crate::analyzer::hir::{
     CoreData, Expr, Goal, GroupId, Literal, LogicalOp, MatchArm, Materializable, Operator, Pattern,
     PhysicalOp, Value,
 };
-use crate::engine::{Continuation, Engine, generator::Generator};
+use crate::engine::{Continuation, Engine, Generator};
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
